guidyapp/requirements.txt
Django==2.0.2
django-node==4.0.0
django-react==0.11.1
django-rest-framework==0.1.0
django-webpack==3.1.0
djangorestframework==3.7.7
optional-django==0.1.0
pytz==2017.3
